Project done

+++ Operating_systems_class/programming_assignment_1/main.c
@@ -12,23 +12,26 @@
 #define MAX_PROCESSES 10
 
 // Function prototypes
-// *Processes started
+// Core functions
 void listCurrentProcesses(pid_t arr[], int num);
 void create_processes(int n);
 void kill_all_child_processes();
 void kill_process(int process_num);
 void resume_process(int process_num);
+void alarm_handler(int signum);
+void sig_handle();
+void r_from_r(int process_num);
 
 // Helper functions throughout the program
-void sig_handle();
 int findIndex(pid_t arr[], int size, int value);
 int stringToNumber(char command[]);
+int stringToNumberSRRQ(char command[]);
+
 // This object is used to manage the states of all the processes, allows for easy look up of each process
 struct state
 {
     bool ready;
     bool running;
-    bool waiting;
     bool terminated;
 };
 
@@ -37,6 +40,23 @@ pid_t process_pids[MAX_PROCESSES] = {0};   // Array to store process IDs, initia
 int num_processes = 0;                     // Counter to keep track of the processes
 struct state processStates[MAX_PROCESSES]; // An array of state objects made so every process has their own state,
 bool waiting = false;
+// For Cn command
+bool is_a_process_running = false;
+
+// For managing run all command
+int current = 0;
+bool rall_pressed = false;
+bool is_running = false;
+
+// Flags for the 2 modes
+bool rrq = false;
+bool fcfs = true;
+
+// For the globally running process
+int global_process_num;
+
+// Number for alarm()
+int numForS;
 
 int main()
 {
@@ -51,10 +71,10 @@ int main()
 
     // Handles ctrl+c
     signal(SIGINT, sig_handle);
+    signal(SIGALRM, alarm_handler);
     while (1)
     {
-
-        // Starting declarations
+        // Initial start up
         printf("shell 5500>>>");
         scanf("%19s", command);
         scanf("%c", &newline);
@@ -91,18 +111,37 @@ int main()
         // Case to create processes
         case 'c':
 
-            numForC = stringToNumber(command);
-            if (numForC >= 5 && numForC <= 10)
+            // If a child process is running, do not create new processes
+            for (int i = 0; i < MAX_PROCESSES; i++)
             {
-                num_processes = 0;
-                create_processes(numForC);
+                if (processStates[i].running)
+                {
+                    is_a_process_running = true;
+                }
             }
-            else
+            // If no process is running, create processes
+            if (!is_a_process_running)
             {
-                printf("Invalid number of processes. Please enter a number between 5 and 10.\n");
-            }
+                numForC = stringToNumber(command);
+                if (numForC >= 5 && numForC <= 10)
+                {
+                    num_processes = 0;
+                    create_processes(numForC);
+                    printf("%d processes made\n", numForC);
+                }
+                else
+                {
+                    printf("Invalid number of processes. Please enter a number between 5 and 10.\n");
+                }
 
-            break;
+                break;
+            }
+            else if (is_a_process_running)
+            {
+                printf("Switch back to shell to make more processes\n");
+                is_a_process_running = false;
+                break;
+            }
 
         case 'l':
             // Error handling
@@ -127,23 +166,80 @@ int main()
                 printf("Invalid command\n");
                 break;
             }
+            if (strlen(command) == 1 && command[0] == 'k')
+            {
+                printf("Please enter an ID number\n");
+                break;
+            }
             numForK = stringToNumber(command);
             kill_process(numForK);
             break;
 
         case 'r':
+
+            // Tells program rall is pressed
+            if (strcmp(command, "rall") == 0)
+            {
+                current = 0;
+                rall_pressed = true;
+                if (global_process_num != 0)
+                {
+
+                    r_from_r(global_process_num);
+                }
+                resume_process(process_pids[current]);
+                break;
+            }
+
             // Error handling
-            if (strlen(command) > 1 && !(isdigit(command[1])))
+            else if (strlen(command) > 1 && !(isdigit(command[1])))
             {
                 printf("Invalid command\n");
                 break;
             }
+            // Error handling
+            else if (strlen(command) == 1 && command[0] == 'r')
+            {
+                printf("Please enter an ID number\n");
+                break;
+            }
             numForR = stringToNumber(command);
+
+            // When a process is runnning and you run another one, this code block handles that logic
+            if (findIndex(process_pids, MAX_PROCESSES, numForR) != -1 && is_running)
+            {
+                r_from_r(global_process_num);
+                if (rall_pressed)
+                {
+                    current = findIndex(process_pids, MAX_PROCESSES, numForR);
+                }
+                resume_process(numForR);
+                break;
+            }
+
             resume_process(numForR);
-            // Pause the while loop
-            while (waiting)
-                ;
             break;
+        case 's':
+            if (strcmp(command, "sfcfs") == 0)
+            {
+                printf("Mode set to first come first serve\n");
+                fcfs = true;
+                alarm(0);
+                break;
+            }
+            else if (strlen(command) > 3 && command[1] == 'r' && command[2] == 'r')
+            {
+                numForS = stringToNumberSRRQ(command);
+                alarm(numForS);
+                printf("Alarm set for %d seconds\n", numForS);
+                rrq = true;
+                break;
+            }
+            else
+            {
+                printf("Invalid command\n");
+                break;
+            }
         default:
             // User entered an invalid command
             printf("Invalid command\n");
@@ -188,7 +284,6 @@ void create_processes(int n)
             process_pids[i] = pid; // Store the PID in the array to be able to use
             processStates[i].ready = true;
             processStates[i].running = false;
-            processStates[i].waiting = false;
             processStates[i].terminated = false;
             num_processes++; // Increment the number of created processes
 
@@ -197,7 +292,15 @@ void create_processes(int n)
         }
     }
 }
-
+void alarm_handler(int signum)
+{
+    // For running processes while the alarm goes off
+    if (is_running)
+    {
+        sig_handle();
+    }
+    alarm(numForS);
+}
 void listCurrentProcesses(pid_t arr[], int size)
 {
     printf("Process IDs:\n");
@@ -212,10 +315,7 @@ void listCurrentProcesses(pid_t arr[], int size)
         {
             printf("Process %d, %d is in the running state\n", i + 1, arr[i]);
         }
-        else if (processStates[i].waiting)
-        {
-            printf("Process %d, %d is in the waiting state\n", i + 1, arr[i]);
-        }
+
         else if (processStates[i].terminated)
         {
             printf("Process %d, %d is now terminated\n", i + 1, arr[i]);
@@ -228,7 +328,9 @@ void kill_process(int process_num)
 {
     // If the terminate is successful, update the values of the process
     int exist = findIndex(process_pids, MAX_PROCESSES, process_num);
-    if (exist != -1)
+
+    // Conditionals to check for a successful terminate
+    if (exist != -1 && (processStates[exist].ready || processStates[exist].running))
     {
         if (kill(process_num, SIGKILL) == 0)
         {
@@ -236,9 +338,11 @@ void kill_process(int process_num)
             int index = findIndex(process_pids, MAX_PROCESSES, process_num);
             processStates[index].ready = false;
             processStates[index].running = false;
-            processStates[index].waiting = false;
             processStates[index].terminated = true;
+
+            is_running = false;
         }
+        // Error handling
         else
         {
             printf("Error: Unable to kill process %d, process does not exist\n", process_num);
@@ -249,33 +353,89 @@ void kill_process(int process_num)
         printf("Error: Unable to kill process %d, process does not exist\n", process_num);
     }
 }
+void r_from_r(int process_num)
+{
+    // If the terminate is successful, update the values of the process
+    int exist = findIndex(process_pids, MAX_PROCESSES, process_num);
+
+    // Conditionals to check for a successful terminate
+    if (exist != -1 && processStates[exist].running)
+    {
+        if (kill(process_num, SIGSTOP) == 0)
+        {
+            int index = findIndex(process_pids, MAX_PROCESSES, process_num);
+            processStates[index].ready = true;
+            processStates[index].running = false;
+            processStates[index].terminated = false;
+            // No process is running right now set reset flags
+            global_process_num = 0;
+            is_running = false;
+            // If rall is pressed, go back to the first element to rereun processes
+            if (rall_pressed)
+            {
+                current = 0;
+            }
+        }
+        // Error handling
+
+        else
+        {
+            printf("Error\n");
+        }
+    }
+    else
+    {
+        printf("Error: Unable to find process %d\n", process_num);
+    }
+}
 
 void resume_process(int process_num)
 {
     // Resumes process
     int exist = findIndex(process_pids, MAX_PROCESSES, process_num);
-    if (exist != -1)
+    // If the process exists and is also not terminated
+    if (exist != -1 && !(processStates[exist].terminated))
     {
         if (kill(process_num, SIGCONT) == 0)
         {
+            global_process_num = process_num;
             printf("Process ID %d resumed and state is now running.\n", process_num);
             int index = findIndex(process_pids, MAX_PROCESSES, process_num);
             processStates[index].ready = false;
             processStates[index].running = true;
-            processStates[index].waiting = false;
             processStates[index].terminated = false;
-            waiting = true;
+
+            // Logic for when all processes are scheduled to run
+            is_running = true;
+            if (rall_pressed)
+            {
+                current++;
+
+                if (current == num_processes)
+                {
+                    current = 0;
+                    rall_pressed = false;
+                }
+            }
         }
         else
         {
             printf("Error: Unable to resume process %d, process does not exist\n", process_num);
         }
     }
+    // When all processes are running but a process in the queue is terminated, use recursion to skip it
+    else if (rall_pressed)
+    {
+        current = current + 1;
+        // Recursion to skip terminated processes with the schedule
+        resume_process(process_pids[current]);
+    }
     else
     {
         printf("Error: Unable to resume process %d, process does not exist\n", process_num);
     }
 }
+
 // Helper function for ctrl + c override
 void sig_handle()
 {
@@ -283,15 +443,24 @@ void sig_handle()
     for (int i = 0; i < MAX_PROCESSES; i++)
     {
         // If the process is running, then turn it back into the ready state
+
         if (processStates[i].running)
         {
             processStates[i].running = false;
             processStates[i].ready = true;
             kill(process_pids[i], SIGSTOP);
+
+            // Keeps track of the current process
+            is_running = false;
+            global_process_num = 0;
+            break;
         }
     }
-    // Set program to continue while loop
-    waiting = false;
+    // For managing rall
+    if (rall_pressed)
+    {
+        resume_process(process_pids[current]);
+    }
 }
 // Helper function through out the program
 int findIndex(pid_t arr[], int size, int value)
@@ -309,10 +478,37 @@ int findIndex(pid_t arr[], int size, int value)
 // Helper function in switch statement
 int stringToNumber(char command[])
 {
+    // Example length of string
     char number_str[20];
     int j = 0;
 
     for (int i = 1; i < strlen(command); i++)
+    {
+        // If the character is a digit, add it to the number string
+        if (isdigit(command[i]))
+        {
+            number_str[j++] = command[i];
+        }
+        else if (j > 0)
+        {
+            // If it's not a digit but we've already started recording digits
+            // It means we've reached the end of the number
+            break;
+        }
+    }
+
+    number_str[j] = '\0'; // Null-terminate the number string
+    // Convert the extracted string number to an integer using atoi()
+    int number = atoi(number_str);
+    return number;
+}
+int stringToNumberSRRQ(char command[])
+{
+    // Example length of string
+    char number_str[20];
+    int j = 0;
+
+    for (int i = 3; i < strlen(command); i++)
     {
         // If the character is a digit, add it to the number string
         if (isdigit(command[i]))
